Eliminate the need for XrApp native module on iOS.
XrApp was used to bridge the ARSession object from the
magic-script-components-react-native library to the
react-native-xr-client library. It still required that the app
developer sleep for a bit (1 second in the sample) before calling
XrApp.shareSession, to ensure that the ARSession had already been
initialized.

This change eliminates the need to call sleep and the need for a
bridging native module. Instead, when XrClientBridge.connect is
called, the iOS implementation will traverse the UIView hierarchy
until it finds an ARSCNView (from which it can obtain the ARSession).
It will poll for up to 30 seconds, sleeping for 200ms in between.